+++ empire/server/common/helpers.py
@@ -33,13 +33,11 @@ Includes:
     complete_path() - helper to tab-complete file paths
     dict_factory() - helper that returns the SQLite query results as a dictionary
     KThread() - a subclass of threading.Thread, with a kill() method
-    slackMessage() - send notifications to the Slack API
 """
 
 import base64
 import binascii
 import ipaddress
-import json
 import logging
 import os
 import random
@@ -48,9 +46,6 @@ import socket
 import string
 import sys
 import threading
-import urllib.error
-import urllib.parse
-import urllib.request
 from datetime import datetime
 
 import netifaces
@@ -778,9 +773,3 @@ class KThread(threading.Thread):
 
     def kill(self):
         self.killed = True
-
-
-def slackMessage(slack_webhook_url, slack_text):
-    message = {"text": slack_text}
-    req = urllib.request.Request(slack_webhook_url, json.dumps(message).encode("UTF-8"))
-    urllib.request.urlopen(req)