+++ empire/server/listeners/dbx.py
@@ -116,11 +116,6 @@ class Listener:
                 "Required": False,
                 "Value": "",
             },
-            "SlackURL": {
-                "Description": "Your Slack Incoming Webhook URL to communicate with your Slack instance.",
-                "Required": False,
-                "Value": "",
-            },
         }
 
         # required: